Little changes in projects.

+++ src/game/ctt/resources/sounds/ogg/OGGsoundLoader.h
@@ -22,7 +22,7 @@ namespace OGG
 	class SoundLoader : public ::SoundLoader
 	{
 	private:
-		static uint32 fileRead(void* buffer, uint32 size, uint32 n, void* opaque);
+		static size_t fileRead(void* buffer, size_t size, size_t n, void* opaque);
 		static int    fileSeek(void* opaque, ogg_int64_t offset, int whence);
 		static int    fileClose(void* opaque);
 		static long   fileTell(void* opaque);
